requirements.txt
openai==1.55.3
httpx==0.27.2
typing-extensions>=4.11.0,<5.0.0
EbookLib
beautifulsoup4
PyMuPDF
lxml
python-docx